requirements.txt
pdf2docx==0.5.1
python-docx==0.8.11
pillow>=9.0.0
pymupdf>=1.23.0
